Get GitHub Repositories by Topic

Get GitHub Repositories by Topic

Usage

get_github_by_topic(topics, token = NULL, limit = 30)

Arguments

topics

A vector of GitHub topics to search for.

token

A GitHub personal access token. If not provided, the function will run without authentication.

limit

The maximum number of results to return. Defaults to 30.

Value

A data frame containing the results of the search query.

Examples

topics <- c("u24ca289073")
df = get_github_by_topic(topics, limit = 50)
head(df)
#> # A tibble: 6 × 11
#>   name       owner stars forks open_issues language license created_at pushed_at
#>   <chr>      <chr> <dbl> <dbl>       <dbl> <chr>    <chr>   <chr>      <chr>    
#> 1 awesome-c… sean…   311    73           1 NA       MIT Li… 2016-07-0… 2025-03-…
#> 2 MultiAssa… wald…    72    35           1 R        NA      2013-07-2… 2025-05-…
#> 3 BiocParal… Bioc…    67    31          31 R        Other   2012-11-1… 2025-05-…
#> 4 SpatialEx… drig…    62    21          33 R        NA      2020-02-1… 2025-05-…
#> 5 HGNChelper wald…    58    10           4 R        NA      2018-07-0… 2024-11-…
#> 6 curatedTC… wald…    44     9           7 R        NA      2016-10-0… 2025-05-…
#> # ℹ 2 more variables: updated_at <chr>, html_url <chr>
dplyr::glimpse(df)
#> Rows: 48
#> Columns: 11
#> $ name        <chr> "awesome-cancer-variant-resources", "MultiAssayExperiment"…
#> $ owner       <chr> "seandavi", "waldronlab", "Bioconductor", "drighelli", "wa…
#> $ stars       <dbl> 311, 72, 67, 62, 58, 44, 33, 27, 27, 21, 20, 20, 19, 17, 1…
#> $ forks       <dbl> 73, 35, 31, 21, 10, 9, 14, 7, 11, 13, 11, 4, 16, 9, 8, 4, …
#> $ open_issues <dbl> 1, 1, 31, 33, 4, 7, 5, 1, 19, 6, 2, 0, 6, 1, 6, 2, 2, 9, 5…
#> $ language    <chr> NA, "R", "R", "R", "R", "R", "R", "R", "R", "R", "Dockerfi…
#> $ license     <chr> "MIT License", NA, "Other", NA, NA, NA, NA, NA, NA, "MIT L…
#> $ created_at  <chr> "2016-07-07T13:16:14Z", "2013-07-22T03:29:27Z", "2012-11-1…
#> $ pushed_at   <chr> "2025-03-06T23:04:53Z", "2025-05-06T15:45:16Z", "2025-05-2…
#> $ updated_at  <chr> "2025-05-06T16:45:16Z", "2025-05-06T15:45:20Z", "2025-05-1…
#> $ html_url    <chr> "https://github.com/seandavi/awesome-cancer-variant-resour…